Added more eclipse files
