+++ eclipse/com.sun.jna.platform/.settings/org.eclipse.pde.core.prefs
