+++ eclipse/com.sun.jna.platform/pom.xml
@@ -75,11 +75,10 @@ com.sun.jna.platform.mac,
 com.sun.jna.platform.unix,
 com.sun.jna.platform.win32,
 com.sun.jna.platform.win32.COM,
-com.sun.jna.platform.win32.COM.proxy,
-com.sun.jna.platform.win32.COM.proxy.annotation,
 com.sun.jna.platform.win32.COM.tlb,
 com.sun.jna.platform.win32.COM.tlb.imp,
 com.sun.jna.platform.win32.COM.util,
+com.sun.jna.platform.win32.COM.util.annotation,
 com.sun.jna.platform.win32.COM.wince
 						</Export-Package>
 					</instructions>
